FEE-709 Improve performance for large data sets in list box. Performed the following optimizations:
* Reduced the number of watchers in the list box by a significant amount
* Moved the search filter into the directive so it isn't calculated on every digest
* Batched the i18n calls so less digest loops are made when rendering the listbox
* Added a GUID to each row so that we can use it in a "track by" clause as part of "ng-repeat" (this allows angular to re-use DOM nodes when possible)

However, the core issue still remains that many DOM nodes are inserted and removed in the page so performance still isn't great. I'd recommend not using more than 500 rows (although 1000 seems to perform okay with ~2 seconds to load).
If you want to calculate the number of watches that will be created on the page, it goes by the following formula:

24 + numRows * (2 + 2 * numCols)